feat: fix reading "broken" Hyper-V DMI data
The fix goes into the `digitalocean/go-smbios` library which seems to be
no longer supported now. So I put as `internal/` sub-package so we can
make easy changes to it in sync with our library.

License check was disabled so that it doesn't complain about Apache
licensed files imported.

Adds support for and testing of extracting HyperV SMBIOS data.

Signed-off-by: Tim Jones <tim.jones@siderolabs.com>
Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>